client: modularized after oauth step for ease of access
